+++ test/opf.jl
@@ -59,7 +59,7 @@
         result = run_ac_opf("../test/data/matpower/case6.m", ipopt_solver)
 
         @test result["status"] == :LocalOptimal
-        @test isapprox(result["objective"], 11567; atol = 1e0)
+        @test isapprox(result["objective"], 11625.3; atol = 1e0)
         @test isapprox(result["solution"]["bus"]["1"]["va"], 0.0; atol = 1e-4)
         @test isapprox(result["solution"]["bus"]["4"]["va"], 0.0; atol = 1e-4)
     end
@@ -111,7 +111,7 @@ end
         result = run_opf("../test/data/matpower/case6.m", ACRPowerModel, ipopt_solver)
 
         @test result["status"] == :LocalOptimal
-        @test isapprox(result["objective"], 11567; atol = 1e0)
+        @test isapprox(result["objective"], 11625.3; atol = 1e0)
         @test isapprox(result["solution"]["bus"]["1"]["va"], 0.0; atol = 1e-4)
         @test isapprox(result["solution"]["bus"]["4"]["va"], 0.0; atol = 1e-4)
     end
@@ -159,7 +159,7 @@ end
         result = run_opf("../test/data/matpower/case6.m", ACTPowerModel, ipopt_solver)
 
         @test result["status"] == :LocalOptimal
-        @test isapprox(result["objective"], 11567; atol = 1e0)
+        @test isapprox(result["objective"], 11625.3; atol = 1e0)
         @test isapprox(result["solution"]["bus"]["1"]["va"], 0.0; atol = 1e-4)
         @test isapprox(result["solution"]["bus"]["4"]["va"], 0.0; atol = 1e-4)
     end
@@ -207,7 +207,7 @@ end
         result = run_dc_opf("../test/data/matpower/case6.m", ipopt_solver)
 
         @test result["status"] == :LocalOptimal
-        @test isapprox(result["objective"], 11396; atol = 1e0)
+        @test isapprox(result["objective"], 11391.8; atol = 1e0)
         @test isapprox(result["solution"]["bus"]["1"]["va"], 0.0; atol = 1e-4)
         @test isapprox(result["solution"]["bus"]["4"]["va"], 0.0; atol = 1e-4)
     end
@@ -218,6 +218,51 @@ end
     #    @test result["status"] == :LocalOptimal
     #    @test isapprox(result["objective"], 79804; atol = 1e0)
     #end
+end
+
+@testset "test nfa opf" begin
+    @testset "3-bus case" begin
+        result = run_opf("../test/data/matpower/case3.m", NFAPowerModel, ipopt_solver)
+
+        @test result["status"] == :LocalOptimal
+        @test isapprox(result["objective"], 5638.97; atol = 1e0)
+    end
+    @testset "5-bus asymmetric case" begin
+        result = run_opf("../test/data/matpower/case5_asym.m", NFAPowerModel, ipopt_solver)
+
+        @test result["status"] == :LocalOptimal
+        @test isapprox(result["objective"], 14810.0; atol = 1e0)
+    end
+    @testset "5-bus gap case" begin
+        result = run_opf("../test/data/matpower/case5_gap.m", NFAPowerModel, ipopt_solver)
+
+        @test result["status"] == :LocalOptimal
+        @test isapprox(result["objective"], -27410.0; atol = 1e0)
+    end
+    @testset "5-bus with asymmetric line charge" begin
+        result = run_opf("../test/data/pti/case5_alc.raw", NFAPowerModel, ipopt_solver)
+
+        @test result["status"] == :LocalOptimal
+        @test isapprox(result["objective"], 1000.0; atol = 1e0)
+    end
+    @testset "5-bus with pwl costs" begin
+        result = run_opf("../test/data/matpower/case5_pwlc.m", NFAPowerModel, ipopt_solver)
+
+        @test result["status"] == :LocalOptimal
+        @test isapprox(result["objective"], 42565.8; atol = 1e0)
+    end
+    @testset "6-bus case" begin
+        result = run_opf("../test/data/matpower/case6.m", NFAPowerModel, ipopt_solver)
+
+        @test result["status"] == :LocalOptimal
+        @test isapprox(result["objective"], 11277.9; atol = 1e0)
+    end
+    @testset "24-bus rts case" begin
+        result = run_opf("../test/data/matpower/case24.m", NFAPowerModel, ipopt_solver)
+
+        @test result["status"] == :LocalOptimal
+        @test isapprox(result["objective"], 61001.2; atol = 1e0)
+    end
 end
 
 @testset "test dc+ll opf" begin
@@ -255,7 +300,7 @@ end
         result = run_opf("../test/data/matpower/case6.m", DCPLLPowerModel, ipopt_solver)
 
         @test result["status"] == :LocalOptimal
-        @test isapprox(result["objective"], 11515; atol = 1e0)
+        @test isapprox(result["objective"], 11574.3; atol = 1e0)
         @test isapprox(result["solution"]["bus"]["1"]["va"], 0.0; atol = 1e-4)
         @test isapprox(result["solution"]["bus"]["4"]["va"], 0.0; atol = 1e-4)
     end
@@ -315,7 +360,7 @@ end
         result = run_opf("../test/data/matpower/case6.m", SOCWRPowerModel, ipopt_solver)
 
         @test result["status"] == :LocalOptimal
-        @test isapprox(result["objective"], 11560; atol = 1e0)
+        @test isapprox(result["objective"], 11472.3; atol = 1e0)
     end
     @testset "24-bus rts case" begin
         result = run_opf("../test/data/matpower/case24.m", SOCWRPowerModel, ipopt_solver)
@@ -425,7 +470,7 @@ end
         result = run_opf_bf("../test/data/matpower/case6.m", SOCBFPowerModel, ipopt_solver)
 
         @test result["status"] == :LocalOptimal
-        @test isapprox(result["objective"], 11567.1; atol = 1e0)
+        @test isapprox(result["objective"], 11472.3; atol = 1e0)
     end
     @testset "24-bus rts case" begin
         result = run_opf_bf("../test/data/matpower/case24.m", SOCBFPowerModel, ipopt_solver)
@@ -498,7 +543,7 @@ end
         result = run_opf("../test/data/matpower/case6.m", QCWRPowerModel, ipopt_solver)
 
         @test result["status"] == :LocalOptimal
-        @test isapprox(result["objective"], 11567; atol = 1e0)
+        @test isapprox(result["objective"], 11484.2; atol = 1e0)
         @test isapprox(result["solution"]["bus"]["1"]["va"], 0.0; atol = 1e-4)
         @test isapprox(result["solution"]["bus"]["4"]["va"], 0.0; atol = 1e-4)
     end
@@ -533,7 +578,7 @@ end
         result = run_opf("../test/data/matpower/case6.m", QCWRTriPowerModel, ipopt_solver)
 
         @test result["status"] == :LocalOptimal
-        @test isapprox(result["objective"], 11567.1; atol = 1e0)
+        @test isapprox(result["objective"], 11512.9; atol = 1e0)
         @test isapprox(result["solution"]["bus"]["1"]["va"], 0.0; atol = 1e-4)
         @test isapprox(result["solution"]["bus"]["4"]["va"], 0.0; atol = 1e-4)
     end
@@ -588,7 +633,7 @@ end
         result = run_opf("../test/data/matpower/case6.m", SDPWRMPowerModel, scs_solver)
 
         @test result["status"] == :Optimal
-        @test isapprox(result["objective"], 11558.5; atol = 1e0)
+        @test isapprox(result["objective"], 11578.8; atol = 1e0)
     end
     # TODO replace this with smaller case, way too slow for unit testing
     #@testset "24-bus rts case" begin